Module wasmer_middleware_common::metering

Structs

ExecutionLimitExceededError

Metering
Metering is a compiler middleware that calculates the cost of WebAssembly instructions at compile time and will count the cost of executed instructions at runtime. Within the Metering functionality, this instruction cost is called points.

Functions

get_points_used
Returns the number of points used by an Instance.

get_points_used_ctx
Returns the number of points used in a Ctx.

set_points_used
Sets the number of points used by an Instance.

set_points_used_ctx
Sets the number of points used in a Ctx.